+++ pages/professeurs/ancien_dossier/ancien_professeurs.module.scss
@@ -1,0 +1,7 @@
+.conteneur {
+  color: white;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
