+++ pages/professeurs/grille/grille.module.scss
@@ -1,0 +1,3 @@
+.conteneur {
+    color: white;
+}
