+++ pages/professeurs/professeurs.module.scss
@@ -1,7 +1,6 @@
 .conteneur {
-  color: white;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+    color: white;
+    :visited, :link {
+        color: green;
+    }
 }
